requirements.txt
asyncio==3.4.3
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudflare==2.6.5
cryptography==2.9.2
Flask==1.1.2
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
jsonlines==1.2.0
jsonpickle==1.4.1
MarkupSafe==1.1.1
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.3.0
python-digitalocean==1.15.0
PyYAML==5.3.1
requests==2.23.0
scp==0.13.2
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
